requirements.txt
# Core dependencies
numpy==1.24.3  # Last stable version for Python 3.11 that's compatible with opencv
opencv-python-headless==4.8.1.78
ultralytics==8.3.59
Pillow==10.4.0

# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==11.0.3

# Already provided by Raspberry Pi OS
# picamera2 - installed via apt as python3-picamera2